requirements.txt
asgiref==3.4.1
Django==3.0.5
django-cors-headers==3.10.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
numpy==1.21.3
psycopg2-binary==2.9.1
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
